pyproject.toml
[project]
name = "aic-dashboard"
version = "0.1.0"
description = "Interactive speech enhancement demo using ai-coustics SDK and Dash"
requires-python = ">=3.10"
dependencies = [
    "dash>=2.17.0",
    "plotly>=5.22.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "pyfar>=0.6.0",
    "aic-sdk",
    "python-dotenv",
    "gunicorn",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "ruff>=0.4.0",
]